+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [Pre-release]
 
 ## [1.2.0] - 2021-08-10
 ### Added
@@ -34,7 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.0] - 2020-05-15
 *initial release*
 
-[Unreleased]: https://github.com/nekr0z/webmention.io-backup/compare/v1.2.0...HEAD
+[Pre-release]: https://github.com/nekr0z/webmention.io-backup/releases/tag/latest
 [1.2.0]: https://github.com/nekr0z/webmention.io-backup/releases/tag/v1.2.0
 [1.1.0]: https://github.com/nekr0z/webmention.io-backup/releases/tag/v1.1.0
 [1.0.2]: https://github.com/nekr0z/webmention.io-backup/releases/tag/v1.0.2