+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Pre-release]
+### Added
+* option to save webmentions for separate pages to separate files in the directory structure that represents the site structure
 
 ## [1.2.0] - 2021-08-10
 ### Added