+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Pre-release]
+
+## [1.3.0] - 2021-08-13
 ### Added
 * option to save webmentions for separate pages to separate files in the directory structure that represents the site structure
 * option to use timestamp to avoid re-downloading old webmentions in SSG mode
@@ -38,6 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 *initial release*
 
 [Pre-release]: https://github.com/nekr0z/webmention.io-backup/releases/tag/latest
+[1.3.0]: https://github.com/nekr0z/webmention.io-backup/releases/tag/v1.3.0
 [1.2.0]: https://github.com/nekr0z/webmention.io-backup/releases/tag/v1.2.0
 [1.1.0]: https://github.com/nekr0z/webmention.io-backup/releases/tag/v1.1.0
 [1.0.2]: https://github.com/nekr0z/webmention.io-backup/releases/tag/v1.0.2